+++ packages/vseed/src/pipeline/spec/chart/pipes/axes/radius.ts
@@ -1,14 +1,8 @@
 import type { IRoseChartSpec } from '@visactor/vchart'
 import type { SpecPipe } from 'src/types'
 
-export const radiusAxis: SpecPipe = (spec, context) => {
+export const radiusAxis: SpecPipe = (spec) => {
   const result = { ...spec } as IRoseChartSpec
-  const { advancedVSeed } = context
-  const { dimensions, measures } = advancedVSeed
-
-  if (!dimensions || !measures) {
-    return result
-  }
 
   if (!result.axes) {
     result.axes = []