+++ packages/vseed/src/pipeline/spec/chart/pipes/axes/yLinearPrimary.ts
@@ -32,20 +32,6 @@ export const yLinearPrimary: SpecPipe = (spec, context) => {
     result.axes = []
   }
 
-  if (!yAxisConfig) {
-    result.axes = [
-      ...result.axes,
-      {
-        visible: true,
-        id,
-        seriesId,
-        type: 'linear',
-        orient: 'left',
-      },
-    ] as ISpec['axes']
-    return result
-  }
-
   const {
     visible = true,
     label,